Down-cast net metadata to reduce bandwidth usage
NetworkObject IDs reduced from ulong to uint
NetVar Delta NetBehaviourIndex reduced from ushort to byte

Rpc bandwidth reduced by 4 bytes, network variable delta updates reduced by 5 bytes per object

+++ com.unity.netcode.gameobjects/Runtime/Core/NetworkBehaviour.cs
@@ -49,7 +49,7 @@ namespace Unity.Netcode
             {
                 Metadata = new RpcMetadata
                 {
-                    NetworkObjectId = NetworkObjectId,
+                    NetworkObjectId = (uint)NetworkObjectId,
                     NetworkBehaviourId = NetworkBehaviourId,
                     NetworkRpcMethodId = rpcMethodId,
                 },
@@ -130,7 +130,7 @@ namespace Unity.Netcode
             {
                 Metadata = new RpcMetadata
                 {
-                    NetworkObjectId = NetworkObjectId,
+                    NetworkObjectId = (uint)NetworkObjectId,
                     NetworkBehaviourId = NetworkBehaviourId,
                     NetworkRpcMethodId = rpcMethodId,
                 },
@@ -632,8 +632,8 @@ namespace Unity.Netcode
                 {
                     var message = new NetworkVariableDeltaMessage
                     {
-                        NetworkObjectId = NetworkObjectId,
-                        NetworkBehaviourIndex = NetworkObject.GetNetworkBehaviourOrderIndex(this),
+                        NetworkObjectId = (uint)NetworkObjectId,
+                        NetworkBehaviourIndex = (byte)NetworkObject.GetNetworkBehaviourOrderIndex(this),
                         NetworkBehaviour = this,
                         TargetClientId = targetClientId,
                         DeliveryMappedNetworkVariableIndex = m_DeliveryMappedNetworkVariableIndices[j]
